+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - replace octocrab with octocrate in client and pull_request modules(pr [#264])
 
+### Changed
+
+- refactor-replace octocrab with octocrate in various modules(pr [#265])
+
 ## [0.1.26] - 2024-08-10
 
 ### Changed
@@ -465,6 +469,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#262]: https://github.com/jerus-org/pcu/pull/262
 [#263]: https://github.com/jerus-org/pcu/pull/263
 [#264]: https://github.com/jerus-org/pcu/pull/264
+[#265]: https://github.com/jerus-org/pcu/pull/265
 [Unreleased]: https://github.com/jerus-org/pcu/compare/v0.1.26...HEAD
 [0.1.26]: https://github.com/jerus-org/pcu/compare/v0.1.25...v0.1.26
 [0.1.25]: https://github.com/jerus-org/pcu/compare/v0.1.24...v0.1.25